[#15] [Android] [UI] Add named argument and rename package
