+++ androidApp/src/main/java/co/nimblehq/avishek/phong/kmmic/android/MainActivity.kt
@@ -5,11 +5,8 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.*
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import co.nimblehq.avishek.phong.kmmic.Greeting
-import co.nimblehq.avishek.phong.kmmic.android.ui.screen.login.SplashScreen
+import co.nimblehq.avishek.phong.kmmic.android.ui.screen.splash.SplashScreen
 import co.nimblehq.avishek.phong.kmmic.android.ui.theme.ApplicationTheme
 
 class MainActivity : ComponentActivity() {
